spike on create_app

+++ deploy/activate.sh
@@ -12,10 +12,13 @@ deploy_directory=/home/$username/deployments
 if [ "$is_laravel" = true ]; then
 
   # .env
-#  if [ ! -f $deploy_directory/.env ]; then
-#      sudo cp $parent_path/.env $deploy_directory/.env
-#      sudo sed -i "s|APP_URL=.*|APP_URL=http://$app_domain|" $deploy_directory/.env
-#  fi
+  if [ ! -f $deploy_directory/.env ]; then
+      echo "-----------------------MISSING .env---------------------------"
+      echo "Create a .env file for the laravel application in the following location:"
+      echo $deploy_directory/symlinks/.env
+      echo "-----------------------MISSING .env---------------------------"
+      exit 1
+  fi
   if [ ! -f .env ]; then
       ln -sf $deploy_directory/symlinks/.env .env
   fi
